Make wrappers public, move them to new namespace
